requirements.txt
jinja2==3.1.4
pyyaml==6.0.2
jsonschema==4.23.0
fastapi==0.115.0
uvicorn==0.30.6